+++ spec/assets/custom-eslint-config/.eslintrc.js
@@ -1,0 +1,6 @@
+module.exports = {
+    "extends": "standard",
+    "rules": {
+        "max-len": ["error", { "code": 200 }]
+    }
+};