+++ spec/nodeApiSpec.js
@@ -44,4 +44,23 @@ describe('node api', () => {
     shell.rm('-rf', 'spec/assets/sample1_output_')
   });
 
+
+
+  it('should accept custom eslintrc', () => {
+    const config = {
+      source: `var foo = { "bar": "This is a bar.", "baz": { "qux": "This is a qux" }, "difficult": "to read" };`,
+      style: 'standard',
+    };
+    const result = formatter(config);
+    expect(result).toContain(`bar: 'This is a bar.',\n  baz: { qux: 'This is a qux' },`)
+
+    const config2 = {
+      source: `var foo = { "bar": "This is a bar.", "baz": { "qux": "This is a qux" }, "difficult": "to read" };`,
+      eslintPath: 'spec/assets/custom-eslint-config/.eslintrc.js',
+    };
+    const result2 = formatter(config2);
+    expect(result2).toContain(`var foo = { bar: 'This is a bar.', baz: { qux: 'This is a qux' }, difficult: 'to read' }`);
+  });
+
+
 });