+++ Changelog.md
@@ -38,3 +38,7 @@
 ## 0.1.0
 
  * input globs, output or re-write support
+
+## 0.1.1
+
+ * custom eslintrc as --style